+++ listen-engine/src/engine/order.rs
@@ -6,18 +6,3 @@ pub struct SwapOrder {
     pub output_token: String,
     pub amount: String,
 }
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PrivyOrder {
-    pub user_id: String,
-    pub address: String,
-    pub caip2: String,
-    pub evm_transaction: Option<serde_json::Value>,
-    pub solana_transaction: Option<String>, // base64
-}
-
-impl PrivyOrder {
-    pub fn is_solana(&self) -> bool {
-        self.caip2.starts_with("solana")
-    }
-}
