+++ listen-engine/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod engine;
 pub mod metrics;
-pub mod privy;
 pub mod redis;
 pub mod server;
 pub use engine::Engine;
