+++ privy/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "privy"
+version = "0.1.0"
+edition = "2021"
+authors = ["piotrostr"]
+license = "MIT"
+
+[dependencies]
+anyhow = "1.0.95"
+jsonwebtoken = "9.3.1"
+serde = { version = "1.0.217", features = ["derive"] }
+reqwest = { version = "0.12.12", features = ["json"] }
+serde_json = "1.0.138"
+thiserror = "2.0.11"
+base64 = "0.22.1"
+tracing = "0.1.41"
+tracing-subscriber = "0.3.19"
